# feat(proto): optimize encoding