--- a/proto/cmd/ch-gen-col/main.tpl
+++ b/proto/cmd/ch-gen-col/main.tpl
@@ -5,9 +5,6 @@ package proto
 
 import (
 	"encoding/binary"
-{{- if .IsFloat }}
-	"math"
-{{- end }}
 )
 
 var _ = binary.LittleEndian // clickHouse uses LittleEndian
@@ -59,34 +56,4 @@ func (c *ColArr) Append{{ .Name }}(data []{{ .ElemType }}) {
   d := c.Data.(*{{ .Type }})
   *d = append(*d, data...)
   c.Offsets = append(c.Offsets, uint64(len(*d)))
-}
-
-// EncodeColumn encodes {{ .Name }} rows to *Buffer.
-func (c {{ .Type }}) EncodeColumn(b *Buffer) {
-  {{- if .Byte }}
-  b.Buf = append(b.Buf, c...)
-  {{- else if .SingleByte }}
-  start := len(b.Buf)
-  b.Buf = append(b.Buf, make([]byte, len(c))...)
-  for i := range c {
-	b.Buf[i + start] = {{ .UnsignedType }}(c[i])
-  }
-  {{- else }}
-  const size = {{ .Bits }} / 8
-  offset := len(b.Buf)
-  b.Buf = append(b.Buf, make([]byte, size * len(c))...)
-  for _, v := range c {
-	{{ .BinPut }}(
-	  b.Buf[offset:offset+size],
-	{{- if .IsFloat }}
-	  math.{{ .Name }}bits(v),
-	{{- else if .Cast }}
-	  {{ .UnsignedType }}(v),
-	{{- else }}
-	  v,
-	{{- end }}
-	)
-	offset += size
-  }
-  {{- end }}
 }