--- a/proto/col_datetime64_gen.go
+++ b/proto/col_datetime64_gen.go
@@ -56,17 +56,3 @@ func (c *ColArr) AppendDateTime64(data []DateTime64) {
 	*d = append(*d, data...)
 	c.Offsets = append(c.Offsets, uint64(len(*d)))
 }
-
-// EncodeColumn encodes DateTime64 rows to *Buffer.
-func (c ColDateTime64) EncodeColumn(b *Buffer) {
-	const size = 64 / 8
-	offset := len(b.Buf)
-	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
-	for _, v := range c {
-		binary.LittleEndian.PutUint64(
-			b.Buf[offset:offset+size],
-			uint64(v),
-		)
-		offset += size
-	}
-}