--- a/proto/col_datetime_unsafe_gen.go
+++ b/proto/col_datetime_unsafe_gen.go
@@ -25,3 +25,20 @@ func (c *ColDateTime) DecodeColumn(r *Reader, rows int) error {
 	}
 	return nil
 }
+
+// EncodeColumn encodes DateTime rows to *Buffer.
+func (c ColDateTime) EncodeColumn(b *Buffer) {
+	if len(c) == 0 {
+		return
+	}
+	offset := len(b.Buf)
+	const size = 32 / 8
+	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
+
+	s := *(*slice)(unsafe.Pointer(&c))
+	s.Len *= 4
+	s.Cap *= 4
+	src := *(*[]byte)(unsafe.Pointer(&s))
+	dst := b.Buf[offset:]
+	copy(dst, src)
+}