--- a/proto/col_decimal128_gen.go
+++ b/proto/col_decimal128_gen.go
@@ -56,17 +56,3 @@ func (c *ColArr) AppendDecimal128(data []Decimal128) {
 	*d = append(*d, data...)
 	c.Offsets = append(c.Offsets, uint64(len(*d)))
 }
-
-// EncodeColumn encodes Decimal128 rows to *Buffer.
-func (c ColDecimal128) EncodeColumn(b *Buffer) {
-	const size = 128 / 8
-	offset := len(b.Buf)
-	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
-	for _, v := range c {
-		binPutUInt128(
-			b.Buf[offset:offset+size],
-			UInt128(v),
-		)
-		offset += size
-	}
-}