--- a/proto/col_decimal128_unsafe_gen.go
+++ b/proto/col_decimal128_unsafe_gen.go
@@ -25,3 +25,20 @@ func (c *ColDecimal128) DecodeColumn(r *Reader, rows int) error {
 	}
 	return nil
 }
+
+// EncodeColumn encodes Decimal128 rows to *Buffer.
+func (c ColDecimal128) EncodeColumn(b *Buffer) {
+	if len(c) == 0 {
+		return
+	}
+	offset := len(b.Buf)
+	const size = 128 / 8
+	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
+
+	s := *(*slice)(unsafe.Pointer(&c))
+	s.Len *= 16
+	s.Cap *= 16
+	src := *(*[]byte)(unsafe.Pointer(&s))
+	dst := b.Buf[offset:]
+	copy(dst, src)
+}