--- a/proto/col_decimal256_safe_gen.go
+++ b/proto/col_decimal256_safe_gen.go
@@ -35,3 +35,17 @@ func (c *ColDecimal256) DecodeColumn(r *Reader, rows int) error {
 	*c = v
 	return nil
 }
+
+// EncodeColumn encodes Decimal256 rows to *Buffer.
+func (c ColDecimal256) EncodeColumn(b *Buffer) {
+	const size = 256 / 8
+	offset := len(b.Buf)
+	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
+	for _, v := range c {
+		binPutUInt256(
+			b.Buf[offset:offset+size],
+			UInt256(v),
+		)
+		offset += size
+	}
+}