--- a/proto/col_enum16_gen.go
+++ b/proto/col_enum16_gen.go
@@ -56,17 +56,3 @@ func (c *ColArr) AppendEnum16(data []Enum16) {
 	*d = append(*d, data...)
 	c.Offsets = append(c.Offsets, uint64(len(*d)))
 }
-
-// EncodeColumn encodes Enum16 rows to *Buffer.
-func (c ColEnum16) EncodeColumn(b *Buffer) {
-	const size = 16 / 8
-	offset := len(b.Buf)
-	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
-	for _, v := range c {
-		binary.LittleEndian.PutUint16(
-			b.Buf[offset:offset+size],
-			uint16(v),
-		)
-		offset += size
-	}
-}