--- a/proto/col_enum8_unsafe_gen.go
+++ b/proto/col_enum8_unsafe_gen.go
@@ -23,3 +23,16 @@ func (c *ColEnum8) DecodeColumn(r *Reader, rows int) error {
 	}
 	return nil
 }
+
+// EncodeColumn encodes Enum8 rows to *Buffer.
+func (c ColEnum8) EncodeColumn(b *Buffer) {
+	if len(c) == 0 {
+		return
+	}
+	offset := len(b.Buf)
+	b.Buf = append(b.Buf, make([]byte, len(c))...)
+	s := *(*slice)(unsafe.Pointer(&c))
+	src := *(*[]byte)(unsafe.Pointer(&s))
+	dst := b.Buf[offset:]
+	copy(dst, src)
+}