--- a/proto/col_float32_gen.go
+++ b/proto/col_float32_gen.go
@@ -4,7 +4,6 @@ package proto
 
 import (
 	"encoding/binary"
-	"math"
 )
 
 var _ = binary.LittleEndian // clickHouse uses LittleEndian
@@ -56,18 +55,4 @@ func (c *ColArr) AppendFloat32(data []float32) {
 	d := c.Data.(*ColFloat32)
 	*d = append(*d, data...)
 	c.Offsets = append(c.Offsets, uint64(len(*d)))
-}
-
-// EncodeColumn encodes Float32 rows to *Buffer.
-func (c ColFloat32) EncodeColumn(b *Buffer) {
-	const size = 32 / 8
-	offset := len(b.Buf)
-	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
-	for _, v := range c {
-		binary.LittleEndian.PutUint32(
-			b.Buf[offset:offset+size],
-			math.Float32bits(v),
-		)
-		offset += size
-	}
 }