--- a/proto/col_float64_gen.go
+++ b/proto/col_float64_gen.go
@@ -4,7 +4,6 @@ package proto
 
 import (
 	"encoding/binary"
-	"math"
 )
 
 var _ = binary.LittleEndian // clickHouse uses LittleEndian
@@ -56,18 +55,4 @@ func (c *ColArr) AppendFloat64(data []float64) {
 	d := c.Data.(*ColFloat64)
 	*d = append(*d, data...)
 	c.Offsets = append(c.Offsets, uint64(len(*d)))
-}
-
-// EncodeColumn encodes Float64 rows to *Buffer.
-func (c ColFloat64) EncodeColumn(b *Buffer) {
-	const size = 64 / 8
-	offset := len(b.Buf)
-	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
-	for _, v := range c {
-		binary.LittleEndian.PutUint64(
-			b.Buf[offset:offset+size],
-			math.Float64bits(v),
-		)
-		offset += size
-	}
 }