--- a/proto/col_float64_safe_gen.go
+++ b/proto/col_float64_safe_gen.go
@@ -36,3 +36,17 @@ func (c *ColFloat64) DecodeColumn(r *Reader, rows int) error {
 	*c = v
 	return nil
 }
+
+// EncodeColumn encodes Float64 rows to *Buffer.
+func (c ColFloat64) EncodeColumn(b *Buffer) {
+	const size = 64 / 8
+	offset := len(b.Buf)
+	b.Buf = append(b.Buf, make([]byte, size*len(c))...)
+	for _, v := range c {
+		binary.LittleEndian.PutUint64(
+			b.Buf[offset:offset+size],
+			math.Float64bits(v),
+		)
+		offset += size
+	}
+}