--- a/proto/col_int8_safe_gen.go
+++ b/proto/col_int8_safe_gen.go
@@ -29,3 +29,12 @@ func (c *ColInt8) DecodeColumn(r *Reader, rows int) error {
 	*c = v
 	return nil
 }
+
+// EncodeColumn encodes Int8 rows to *Buffer.
+func (c ColInt8) EncodeColumn(b *Buffer) {
+	start := len(b.Buf)
+	b.Buf = append(b.Buf, make([]byte, len(c))...)
+	for i := range c {
+		b.Buf[i+start] = uint8(c[i])
+	}
+}